transformation_and_modeling/requirements.txt
pandas==2.2.3
pyarrow==17.0.0
pytest==8.3.2
pyspark==3.5.1
apache-airflow==2.10.2

